requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
sqlalchemy>=2.0.28
psycopg2-binary>=2.9.9
pydantic>=2.6.3
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
# Social & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
# Multimedia
cloudinary>=1.38.0
# Migrations
alembic>=1.13.1
requests>=2.31.0
